+++ lib_comfyui/comfyui/pre_main.py
@@ -7,8 +7,6 @@ import runpy
 from lib_comfyui import (
     custom_extension_injector,
     ipc,
-    ipc_callback,
-    ipc_strategies,
 )
 from lib_comfyui.comfyui import routes_extension, queue_tracker
 from lib_comfyui.webui import paths, settings
@@ -32,9 +30,9 @@ def setup_ipc():
     print('[sd-webui-comfyui]', 'Setting up IPC...')
     ipc_strategy_class_name = os.getenv('SD_WEBUI_COMFYUI_IPC_STRATEGY_CLASS_NAME')
     print('[sd-webui-comfyui]', f'Using inter-process communication strategy: {settings.ipc_display_names[ipc_strategy_class_name]}')
-    ipc_strategy_factory = getattr(ipc_strategies, os.getenv('SD_WEBUI_COMFYUI_IPC_STRATEGY_CLASS_NAME'))
-    ipc.current_callback_listeners = {'comfyui': ipc_callback.CallbackWatcher(ipc.call_fully_qualified, 'comfyui', ipc_strategy_factory)}
-    ipc.current_callback_proxies = {'webui': ipc_callback.CallbackProxy('webui', ipc_strategy_factory)}
+    ipc_strategy_factory = getattr(ipc.strategies, os.getenv('SD_WEBUI_COMFYUI_IPC_STRATEGY_CLASS_NAME'))
+    ipc.current_callback_listeners = {'comfyui': ipc.callback.CallbackWatcher(ipc.call_fully_qualified, 'comfyui', ipc_strategy_factory)}
+    ipc.current_callback_proxies = {'webui': ipc.callback.CallbackProxy('webui', ipc_strategy_factory)}
     ipc.start_callback_listeners()
     atexit.register(ipc.stop_callback_listeners)
